backend/requirements.txt
cryptography==38.0.1
Flask==2.2.3
Flask_Cors==3.0.10
PyJWT==2.9.0
requests==2.28.2
gunicorn==20.1.0
Werkzeug==2.3.6